Implement comparison with nil
This might need to be revisited in the future; for now, the runtime
will need to implement a templated is_nil() function.  For basic
types, it'll be specialized for their respective nil values. For
composite types, they'll define an _isNil_() method to compare all
their members agains their respective nil values.