Borrow escape analysis from gopherjs
This is a far cry from the real deal -- there are many cases that
are not covered.  But it's a start, and easier than porting esc.go
from gc.

Nothing is done differently now; just a comment next to the variable
that's escaping.  When GC is in place, that will be used instead.

Other things that should be considered are the heap-allocation of
things that are "too big" to fit in the stack.  This will require
heuristics similar to the ones used by esc.go, but will need to be
more conservative since stacks on embedded systems are often small.